refactor: registry loader (#111)
